// richmindale-app.client/src/app/shared/models/public/courses-details.model.ts
export class CoursesDetails {
    Id: any;
    ProgramDetailsId: any;
    ProgramId: any;
    Semester: any;
    TermId: any;
    CourseId: any;
    CourseCode: any;
    CourseTitle: any;
    CourseDescription: any;
    LongDescription: any;
    Revision: any;
    RevisionDate: any;
    PreRequisites: any;
    Credits: any;
    Programs: any;
    Tuition: any;
    Miscellaneous: any; 
}