// richmindale-app.client/src/app/shared/models/public/program-details.model.ts
export class ProgramDetails {
    Id: any;
    ProgramId: any;
    ProgramCode: any;
    ProgramTitle: any;
    ProgramDescription: any;
    Revision: any;
    RevisionDate: any;
    EstimatedCost: any;
    TotalCredits: any;
    NoOfCourses: any;
    Semesters: any;
    AdmissionApplicationFee: any;
    TuitionPerUnit: any;
    MiscellaneousPerUnit: any;
    CreditTransferFee: any;
    DocumentRequestFee: any;
    ExpediteFee: any;
}